+++ back-end/src/main/java/com/awsft/knifeandmustache/update_dto/UpdateServiceDTO.java
@@ -1,15 +1,17 @@
 package com.awsft.knifeandmustache.update_dto;
 
+import java.math.BigDecimal;
+
 import com.awsft.knifeandmustache.model.ECategoryService;
 
 public class UpdateServiceDTO {
     private Long id;
     private String serviceDescription;
     private Integer duration;
-    private Double value;
+    private BigDecimal value;
     private ECategoryService serviceCategory;
 
-    public UpdateServiceDTO(Long id, String serviceDescription, Integer duration, Double value, ECategoryService serviceCategory) {
+    public UpdateServiceDTO(Long id, String serviceDescription, Integer duration, BigDecimal value, ECategoryService serviceCategory) {
         this.id = id;
         this.serviceDescription = serviceDescription;
         this.duration = duration;
@@ -35,10 +37,10 @@ public class UpdateServiceDTO {
     public void setDuration(Integer duration) {
         this.duration = duration;
     }
-    public Double getValue() {
+    public BigDecimal getValue() {
         return value;
     }
-    public void setValue(Double value) {
+    public void setValue(BigDecimal value) {
         this.value = value;
     }
     public ECategoryService getServiceCategory() {
